bbc-basic: Specify the file name at the top of each library file.
This means that they can be saved using "SAVE" without a filename, which
will make tokenising them a little less painful.  I haven't done this to
the step files because those have "." in their canonical names, which
turns into "/" on RISC OS, so there's no good name to put in the
comment.