bbc-basic: Fix FNenv_find and FNenv_get so they can return nil properly.
Using FNhashmap_get meant that FNenv_find couldn't tell the difference
between failing to find an entry, and finding an entry containing nil.
Now it uses FNhashmap_contains to check whether the value is present
first.  This makes it much clearer why FNenv_find and FNenv_get are
strctured the way they are.