bbc-basic: Minimal support for vectors (read/print only, deferred from step 1).
A vector is just a list in a silly hat, or more precisely a special flag
in its head cell.  This means that all the list functions work on
vectors as well, but it may be necessary to rename them in order to
avoid confusion, since at least FNis_list needs to be more specific in
order for FNEVAL to work properly.