vala: allow metadata pointer to be null.
My base class Mal.ValWithMetadata defaulted to filling in an object's
metadata pointer with an instance of Mal.Nil. Since Mal.Nil objects
are not all commoned together in this implementation, it seems like a
lot of pointless effort to create all those extra Nil objects, when
most of them will never be consulted at all. So now we initialise the
metadata pointer to null rather than an instance of Nil, and
manufacture a Nil as the return value only when the core 'meta'
function is asked to return the metadata from such an object.

(Also, Mal.Nil itself is a subclass of ValWithMetadata, and to avoid
infinite recursion of constructors, I linked it to itself as its own
metadata, creating a trivial reference cycle. That created quite a big
memory leak all by itself! But since I'm about to introduce a garbage
collector which will solve the leakage problem much more generally,
that's no longer an important reason to make this change.)