bbc-basic: Disallow metadata on string-like types.
Objects that have a direct reference into S$() can't safely be cloned,
because the GC assumes that when such an object is freed, the string
in S$() can be freed as well.  This assumption would be broken by
cloning, so the cloning of such types by FNwith_meta must be
disallowed.

This has the unexpected, but in retrospect obvious, consequence that
single-member hashmaps can't have metadata.  That's obviously a bit
silly, so a more comprehensive solution is probably warranted in due
course.