perl: Change calling convention to use variadic functions.
Till now, the arguments to each function at the mal layer were passed to
the corresponding function at the perl layer as a single List argument.
This necessitated a lot of dereferencing in the core functions and was
also contrary to the process guide which recommends using variadic
functions.

Since perl functions are intrinsically variadic, this change makes the
code somewhat simpler, and enables further simplifications.