tests: remove atom metadata test.
This was a commented out test I added early as something that might be
interesting even though Clojure does not support it. However, I hadn't
thought through it fully. Most objects are cloned when meta-data is
added. It is weird to clone an atom because then you have two
references that can diverge (i.e. they are different atoms but start
out pointing to the same underlying object). But mutating an atom is
the wrong behavior too i.e. with-meta would be a pseudo form of
reset! just for the metadata part.