get-ci-matrix.py: ignore IMPLS.yml and Makefile.impls
This isn't a perfect solution because if somebody pushes a new
implementation and then later pushes the updates to these files, it
probably won't trigger any implementations. However, the workflow can
always be manually triggered for a specific implementation and branch so
that's probably okay. At least this way, the addition of a new
implementation won't always trigger every other implementation to be
tested.